devex(Makefile): add target for copying git hooks
Add a target in the Makefile for copying git hooks that make preparing
commit messages easier and more descriptive.

The target is PHONY, and copies all the hooks from .git-dev/hooks into
.git/hooks, and makes them executable.

To disable the git hooks, all one needs to do is append the `.sample`
suffix to the hook's file name inside `.git/hooks`.

ref: #2256