[xy] Disable autoreload in non-dev environments. (#5136)
# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Disable autoreload in non-dev environments to reduce CPU usage.

Profiled webserver CPU usage with library yappi.
<img width="1356" alt="image"
src="https://github.com/mage-ai/mage-ai/assets/80284865/31933907-b88d-414b-a58b-ae01950d0a45">

Columns in Profiling Output
* ncalls: The number of calls to the function. If this shows two numbers
separated by a slash (e.g., 6155/6794), it means the function was called
recursively. The first number is the number of primitive calls, and the
second number is the total calls.
* tottime: The total time spent in the given function (excluding time
made in calls to sub-functions).
* percall (tottime): The total time divided by the number of calls (the
first number in ncalls). This represents the average time spent in the
function per call.
* cumtime: The cumulative time spent in this function and all
sub-functions (from invocation till exit). It is the sum of the tottime
and the cumulative time of all sub-functions.
* percall (cumtime): The cumulative time divided by the number of calls.
This represents the average time spent in the function and its
sub-functions per call.

autoreload has a relatively high tottime and cumtime, suggesting it
could be a good candidate for optimization.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Tested locally. After disable autoreloading. The CPU usage went
down from 5% to 0.15% when no requests are made.


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->