[Bug] Updated SalesForce source to handle multiple date formats (#5493)
# Description
For Issue #5409 
When configuring the Salesforce Source in a batch pipeline the
'start_date' is being converted before it is passed through - even
though it is correctly entered in the config.

The root cause looks like it is with how yaml.safe_load parses the date
string and returns to the frontend. Opted to change the SalesForce data
integration to handle both formats to reduce impact. Updated to
`strptime_to_utc` as `strptime` is
[deprecated](https://github.com/singer-io/singer-python/blob/master/singer/utils.py#L30C8-L30C18)

Once this bug was resolved, there was an initial error with the sync
because the `raw_state` was not set. Updated to default the raw_state


# How Has This Been Tested?
Verified the integration using the SalesForce data integration within a
batch pipeline locally. Confirmed loading data from SalesForce with
sandbox
https://www.loom.com/share/b460861fb4804ea7b072d726a455b157

- [x] Verified Test integration with date `2024-01-01T00:00:00Z` 
- [x] Verified loading data with date `2024-01-01T00:00:00Z` 
- [x] Verified Test integration with date `2024-09-01T00:00:01+00:00`
- [x]  Verified loading data with date `2024-09-01T00:00:01+00:00`
- [x] Verified test integration FAILED with bad date
`2024-02-30T00:00:00Z`



# Checklist
- [ ] The PR is tagged with proper labels (bug, enhancement, feature,
documentation) -- I am not able to add labels
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation (N/A)

cc: @wangxiaoyou1993