[jk] Allow user to set number of lines displayed for block output sample preview (#5485)
# Description
- The block output sample preview in the Pipeline Editor defaults to 10
lines, but users may want to decrease this amount to improve browser
performance, especially if they have many blocks in their pipeline or
blocks that returns dataframes with many columns / larges values in
those columns.
- The environment variable `DATAFRAME_SAMPLE_COUNT_PREVIEW` is used to
set the number of lines for the block output sample preview.
- Make the "Disable output preview" setting wording clearer. It says
"Disable output preview on page load" now since the block's sample
output is still visible after running a block, but users might think
that the output is disabled at all times. This setting prevents the
block sample output from appearing when the page initially loads. Also
added `disable_output_preview` as a reserved key so it does not appear
in the "Block variables" section of the block settings, since it is part
of a block's config.

![image](https://github.com/user-attachments/assets/a26f53f5-329a-4bf9-835d-78787180ad40)


# How Has This Been Tested?
- Confirmed that the number of rows in the block sample output matches
the value used for the env var `DATAFRAME_SAMPLE_COUNT_PREVIEW`
- Value of `5` used below:

![image](https://github.com/user-attachments/assets/b7ea44fa-5a66-4242-bc9d-2acb52572f8d)

# Checklist
- [X] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->