check_connection/_async times out with lots of branches
> File "/usr/local/lib/python3.10/site-packages/mage_ai/data_preparation/git/__init__.py", line 213, in check_connection
>   await check_connection_async(self.repo.git, self.origin.name)
> File "/usr/local/lib/python3.10/site-packages/mage_ai/data_preparation/git/utils.py", line 220, in check_connection_async
>   await poll_process_with_timeout(
> File "/usr/local/lib/python3.10/site-packages/mage_ai/data_preparation/git/utils.py", line 210, in poll_process_with_timeout
>   raise TimeoutError

we recently migrated a repo to github with a few hundred branches and
~1500 PRs

check_connection always times out with that many branches/PRs as it
fills up the stdout pipe. that in turn breaks git syncing.

using asyncio.subprocess would be ideal, however git.cmd.Git doesn't
seem to have a way for that. plus we are in an async context, calling
proc.communicate() instead of polling/checking proc.poll() would uhh,
not be great either.

fix by draing the stdout pipe while polling and waiting for the command to exit

( in theory stderr could also fill up, that would also signal a more
legitimate error so we ignore that for now. )