feat: Allow to set server_name on sentry logging (#5789)
# Description
My company uses multiple Mage instances depending on the clients or
projects. We want to take advantage of the log send in Sentry, but it is
currently impossible to tell which Mage instance returns logs on the
same Sentry instance. This PR resolves the issue by allowing the server
name to be configured in the env variables when initializing the Sentry
connector.


# How Has This Been Tested?
I have made tests with our local sentry instance.


# Checklist
- [ ] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation